requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.1
